Meadow.Linux: Add support for Raspberry Pi 5 pinout
The Raspberry Pi 5 pinout is quite different from the previous
versions, so we need to add a new pinout class for it. It use a
in house developed silicon for I/O, the RP1 I/O controller. It's
comes with a new pinctrl driver that changed the sysfs base from
0 to 53 and the gpiochip bank from gpiochip0 to gpiochip4.

However, reusing what we already have from previous versions of
Raspberry Pi, we can support this simply by modifying
the LinuxFlexiPin to allow changing the SysFsGpio and
GpiodChip.

Signed-off-by: Matheus Castello <matheus.castello@toradex.com>